% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apidemo.R
\name{ask_qf}
\alias{ask_qf}
\title{Ernie 4,0 API demo}
\usage{
ask_qf(text, API_KEY, SECRET_KEY)
}
\arguments{
\item{text}{Your question}

\item{API_KEY}{Obtained from bce.baidu.com}

\item{SECRET_KEY}{Obtained from bce.baidu.com}
}
\value{
Qianfan LLM responce
}
\description{
Ernie 4,0 API demo
}
