% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apidemo.R
\name{get_access_token}
\alias{get_access_token}
\title{Get token to connect LLM model}
\usage{
get_access_token(API_KEY, SECRET_KEY)
}
\arguments{
\item{API_KEY}{Obtained from bce.baidu.com}

\item{SECRET_KEY}{Obtained from bce.baidu.com}
}
\value{
One series generated by baidu.com
}
\description{
Get token to connect LLM model
}
